==> R/00_start.R <==
## ADD DOCUMENTATION TO YOUR PROJECT ----

## Add meta data about your project to DESCRIPTION
d6:::fill_desc(
  pkg_title = "Stommel_2023_WildlifeBiol",
  pkg_description = "Effect of human induced surface water scarcity on herbivore distribution during the dry season in Ruaha National Park, Tanzania",
  author_first_name = "Stephanie",
  author_last_name = "Kramer-Schadt",
  author_email = "kramer@izw-berlin.de",
  repo_url = NULL
)

## Add and fill the readme
usethis::use_readme_md()

## Add license if needed
## See ?usethis::use_mit_license for more information
usethis::use_mit_license( name = "Your Name" )

## ADD PACKAGE DEPENDENCIES ----

## Add one line by package you want to add as dependency
usethis::use_package("")
